Yigo's CST is now part of the library!!!
git-svn-id: http://svn.recoded.cl/libcds@34 41656f23-c745-496b-8a1c-052032c6b90f

+++ trunk/src/utils/libcdsTrees.h
@@ -17,9 +17,9 @@ namespace cds_utils{
 	 * @param len Length in bits of each field
 	 * @param index Position to be retrieved
 	 * */
-	inline uint get_field_64(uint *A, uint len, uint index) {
+	inline size_t get_field_64(size_t *A, size_t len, size_t index) {
 		if(len==0) return 0;
-		uint result;
+		size_t result;
 		long long i=1, j=1;
 		i=i*index*len/W, j=j*index*len-W*i;
 		if (j+len <= W)
@@ -31,11 +31,11 @@ namespace cds_utils{
 		return result;
 	}
 
-	inline void set_field_64(uint *A, uint len, uint index, uint x) {
+	inline void set_field_64(size_t *A, size_t len, size_t index, size_t x) {
 		if(len==0) return;
 		long long i=1, j=1;	  
 		i= i*index*len/W, j= j*index*len-i*W;
-		uint mask = ((j+len) < W ? ~0u << (j+len) : 0)
+		size_t mask = ((j+len) < W ? ~0u << (j+len) : 0)
 			| ((W-j) < W ? ~0u >> (W-j) : 0);
 		A[i] = (A[i] & mask) | x << j;
 		if (j+len>W) {
@@ -50,9 +50,9 @@ namespace cds_utils{
 	 * @param t2 Position in the text where start the second suffix
 	 * @param lim Maximum value that can be returned
 	 * */
-	inline uint lcp_length(char *text, uint t1, uint t2, uint lim){
-		uint cont=0;
-		for(uint i=0; i<lim;i++){
+	inline size_t lcp_length(char *text, size_t t1, size_t t2, size_t lim){
+		size_t cont=0;
+		for(size_t i=0; i<lim;i++){
 			if(text[t1+i]==text[t2+i])
 				cont++;
 			else
@@ -63,8 +63,8 @@ namespace cds_utils{
 
 
 	/* returns e[p..p+len-1], assuming len <= W*/
-	inline uint bitread (uint *e, uint p, uint len){ 
-		uint answ=0;
+	inline size_t bitread (size_t *e, size_t p, size_t len){ 
+		size_t answ=0;
 		e += p/W; p %= W;
 		answ = *e >> p;
 		if (len == W){ 
@@ -80,7 +80,7 @@ namespace cds_utils{
 	}
 
 	/* writes e[p..p+len-1] = s, len <= W*/
-	inline void bitwrite (register uint *e, register uint p, register uint len, register uint s){ 
+	inline void bitwrite (register size_t *e, register size_t p, register size_t len, register size_t s){ 
 		e += p/W; 
 		p %= W;
 		if (len == W){
